refactor(#218): move api definitions to feature directories
