+++ apps/frontend/src/features/drawing/components/index.ts
@@ -12,6 +12,5 @@ export * from './IndexRow';
 export * from './IndexSection';
 export * from './ConstraintRow';
 export * from './ConstraintSection';
-export * from './Memo';
 export * from './TempMemoPreview';
 export * from './EditableRowBase';
