+++ apps/frontend/src/features/drawing/hooks/index.ts
@@ -8,7 +8,6 @@ export * from './useIndexes';
 export * from './useConstraints';
 export * from './useViewport';
 export * from './useSchemaEditor';
-export * from './useMemoStore';
 export * from './useCanvasInitialization';
 export * from './useCanvasKeyboard';
 export * from './useCanvasNodes';
