+++ apps/frontend/src/features/drawing/index.ts
@@ -1,4 +1,3 @@
 export * from './components';
 export * from './hooks';
 export * from './types';
-export * from './context';
