+++ apps/frontend/src/lib/api/client.ts
@@ -1,6 +1,6 @@
 import axios, { type AxiosError, type InternalAxiosRequestConfig } from 'axios';
 import { authStore } from '../../store/auth.store';
-import { refreshToken } from './auth/api';
+import { refreshToken } from '@/features/auth/api';
 
 const API_BASE_URL: string =
   import.meta.env.VITE_BASE_URL || 'http://localhost:8080/api/v1.0';
