+++ apps/frontend/src/lib/api/index.ts
@@ -1,4 +1,2 @@
 export * from './client';
-export * from './auth';
-export * from './memo';
 export * from './types';
